Improve polymorphism of DeepEvalBaseBenchmark
Currently, every instance of `DeepEvalBaseLLM` is expected to have an `evaluate` api that takes in a required `model` argument and then uses that model to output a dict.

This PR just codifies the interfaces and adds `kwargs` so that a script can call any benchmark with arbitrary kwargs (that may not apply to this benchmark, such as `batch_size`).

The main use case of this is suppose we have a list of benchmarks

```
benchmarks: list[DeepEvalBaseBenchmark] = [TruthfulQA(), SQuAD()]
kwargs = {"batch_size": 3}
chosen_benchmark_idx: int = ...
model = ...

benchmarks[chosen_benchmark_idx].evaluate(model, **kwargs)
```

This will allow us to run `SQuAD`, which does **NOT** have a `batch_size` argument without crashing, and if we choose to run `TruthfulQA`, then the `batch_size` will be parsed in the function.

However, if you were to attempt this with the current code you will encounter the error
```
SQuAD.evaluate() got an unexpected keyword argument 'batch_size'
```